Refactor analyzer Illuminate classes into separate files.
